+++ src-docs/src/views/badge/badge_with_icon.js
@@ -10,6 +10,8 @@ export default () => (
 
     <EuiBadge iconType="check">Default</EuiBadge>
 
+    <EuiBadge iconType="lensApp">Primary with an app icon</EuiBadge>
+
     <EuiBadge iconType="returnKey" />
   </div>
 );